% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prison.R
\name{add_inflows_det_delta_courts}
\alias{add_inflows_det_delta_courts}
\title{Add delta inflows to base inflows}
\usage{
add_inflows_det_delta_courts(inflows, delta)
}
\description{
Add delta inflows to base inflows
}
