% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentencing.R
\name{calculate_inflows_det_delta}
\alias{calculate_inflows_det_delta}
\title{Calculate receptions by sentence band by applying reception rates to a series
of disposal volumes by disposal type.}
\usage{
calculate_inflows_det_delta(cc_disposals, mc_disposals, sentencing_rates)
}
\arguments{
\item{cc_disposals}{A tibble representing a time series of Crown Court
disposal volumes by type.}

\item{mc_disposals}{A tibble representing a time series of magistrates' court
disposal volumes by type.}

\item{sentencing_rates}{A tibble with columns, \code{disposal_type},
\code{senband1}, \code{senband2}, \code{senband3} and \code{senband4},
representing the coefficients for converting disposal types to receptions
by sentence band.}
}
\value{
A table of prison inflows, split by sentence band.
}
\description{
Calculate receptions by sentence band by applying reception rates to a series
of disposal volumes by disposal type.
}
