% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{format_params}
\alias{format_params}
\title{Format parameters.}
\usage{
format_params(params)
}
\arguments{
\item{params}{The parameters passed to \code{run_prisonsreadyreckoner()}.}
}
\value{
The parameters with additional variables not seen by the user and
others formatted for onward use.
}
\description{
Exported only for use by the Shiny app. Not to be used by a user of the
package.
}
