% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_datasets}
\alias{load_datasets}
\title{Load data from S3}
\usage{
load_datasets(params)
}
\arguments{
\item{params}{The parameters passed to \code{run_prisonsreadyreckoner()}.}
}
\value{
A list containing the contents of all files provided in params.
}
\description{
Exported only for use by the Shiny app. Not to be used by a user of the
package.
}
