% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prison.R
\name{make_remand_filter_in}
\alias{make_remand_filter_in}
\title{Build population impact filters for use in remand modelling. Inflow impacts.}
\usage{
make_remand_filter_in(remand_rate, no_bail_rate, ctl, projection_length_months)
}
\arguments{
\item{remand_rate}{The fraction of receipts, to which this filter will be
applied, being remanded in custody.}

\item{no_bail_rate}{The fraction of those remanded in custody who will be
kept in prison after their CTL has exceeded.}

\item{ctl}{The custody time limit in months.}

\item{projection_length_months}{The number of months in the forecast period.}
}
\value{
A tibble representing, for lags starting from zero, the unit remand
population impact of a magistrates' court receipt.
}
\description{
Exported only for use by the Shiny app. Not to be used by a user of the
package.
}
\details{
Function to generate a linear response function (filter) representing the
unit population impact of a receipt in the magistrates' court on remand
population. It is assumed that only a fraction of relevant receipts will
be associated with remand and that some proportion of these will leave remand
after the custody time limit (CTL).
}
