% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prison.R
\name{make_remand_filter_out}
\alias{make_remand_filter_out}
\title{Build population impact filters for use in remand modelling. Outflow impacts.}
\usage{
make_remand_filter_out(
  remand_rate,
  no_bail_rate,
  ctl,
  projection_length_months
)
}
\arguments{
\item{remand_rate}{The fraction of disposals, to which this filter will be
applied, having been remanded in custody at disposal.}

\item{no_bail_rate}{The fraction of those remanded in custody who will be
kept in prison after their CTL has exceeded.}

\item{ctl}{The custody time limit in months.}
}
\value{
A tibble representing, for lags starting from zero, the unit remand
population impact of a Crown Court disposal.
}
\description{
Function to generate a linear response function (filter) representing the
unit population impact of a disposal in the Crown court on remand population.
It is assumed that only a fraction of relevant disposals will be associated
with remand and that some proportion of these will have left remand after the
custody time limit (CTL) anyway. Additionally assuming that the disposal can
occur any time within the individual's CTL, the filter is most negative at a
lag of zero and declines linearly throughout a period equivalent to the CTL.
A steady state is reached, represenying the probability that the individual's
circumstances do not warrant release on bail.
}
