% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojstockr.R
\name{mojstockr_build_stock}
\alias{mojstockr_build_stock}
\title{Build multiple stocks from corresponding inflows and outflows.}
\usage{
mojstockr_build_stock(inflows, outflows, non_data_cols = NULL)
}
\arguments{
\item{inflows}{A data frame of inflows with flows in consecutive time steps
represented by consecutive 'data' columns.}

\item{outflows}{A data frame of outflows with flows in consecutive time steps
represented by consecutive 'data' columns.}

\item{non_data_cols}{A character vector specifying the columns in inflows and
outflows designated as non-data.}
}
\value{
A tibble representing the integration of inflows and outflows for
each factor in the two data frames where factors will be designated by
\code{non_data_cols}
}
\description{
Function to construct a stock as the sum of all inflows minus the sum of all
outflows.
}
\details{
Inflows and outflows must be a data frame of the same general structure,
having a set of non-data columns and a set of data columns. Consecutive data
columns will represent flows at consecutive time steps.

The function works by binding the inflows to minus the outflows, summing
across the two and then integrating across time steps.
}
\examples{
\dontrun{
inflows <- dataframe(sex = c("female", "male"), t0 = rpois(2, 10),
               t1 = rpois(2, 10), t2 = rpois(2, 10), t3 = rpois(2, 10),
               t4 = rpois(2, 10), t5 = rpois(2, 10))
outflows <- dataframe(sex = c("female", "male"), t0 = rep(2, 2),
               t1 = rep(2, 2), t2 = rep(2, 2), t3 = rep(2, 2),
               t4 = rep(2, 2), t5 = rep(2, 2))
stock <- build_stock(inflows, outflows, c("sex"))
}
}
