% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojstockr.R
\name{mojstockr_calculate_proportions}
\alias{mojstockr_calculate_proportions}
\title{Calculate conditional probabilities over multiple variables}
\usage{
mojstockr_calculate_proportions(
  df,
  predicate_columns,
  conditional_columns,
  proportions_column_name = "proportion",
  weights_column = NULL
)
}
\arguments{
\item{df}{String. Dataframe with multiple columns.}

\item{predicate_columns}{String. Columns on which to group data into groups.}

\item{conditional_columns}{String. Columns on which to group data into
subgroups.}

\item{proportions_column_name}{\emph{Optional.} The name of the column
containing the calculated proportions. Defaults to 'proportion'.}

\item{weights_column}{\emph{Optional.} String. Column containing weights to
use in calculation of proportions. Higher values correspond to higher
weights.}
}
\value{
A tibble listing the groups and subgroups, and giving the proportion
of data within each subgroup. Subgroups without any counts are omitted.
}
\description{
Group dataframe rows that have non-unique combinations in specified columns
and, for each group, calculate the proportion of data belonging to specified
subgroups.
}
\details{
\code{mojstockr_calculate_proportions} accepts an input dataframe and groups
row that have non-unique combinations of values in the columns specified by
the argument \code{predicate_columns}. Within each of the resulting groups,
rows are further grouped into subgroups that have non-unique combinations of
values in the columns specified by the argument \code{conditional_columns}.
The proportions of data within each subgroup are then calculated. (Note that
these proportions are calculated relative to the size of the group housing
the subgroup, \emph{not} relative to the size of all data within the whole
dataframe.) By default, proportions are based on the number of rows belonging
to each subgroup. Different rows can be given different weightings using
values in a column specified by the optional argument \emph{weights_column}.
}
\examples{
\dontrun{
df <- data.frame(a = c("France", "France", "France", "France", "Germany",
                       "Germany", "Germany", "Germany", "Poland", "Poland",
                       "Ukraine", "Ukraine", "Ukraine", "Ukraine", "Ukraine"),
                 b = c("Farm", "Farm", "Farm", "Village", "City", "City",
                       "City", "City", "Village", "Town", "Town", "Town",
                       "Town", "Town", "Town"),
                 c = c("male", "female", "unknown", "male", "female",
                       "unknown", "unknown", "unknown", "unknown", "unknown",
                        "male", "female", "female", "female", "female"),
                 d = c("A", "B", "C", "D","E", "F", "G", "H", "A", "B", "C",
                       "D", "E", "F", "G"),
                 volume = c(12, 20, 8, 15, 25, 20, 7, 2, 38, 3, 45, 2, 9, 13,
                            16)
       )

prob <- mojstockr_calculate_proportions(df, c('a', 'b'), c('c'), weights_column = "volume")
}
}
