% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojstockr.R
\name{mojstockr_make_lag_filter}
\alias{mojstockr_make_lag_filter}
\title{Build a lag filter for use in simplified stock-flow modelling}
\usage{
mojstockr_make_lag_filter(lag, N = NULL)
}
\arguments{
\item{lag}{The lag to represent. A real, non-negative number.}

\item{N}{The maximum time step to include in the filter, counting from zero.
If \code{N+1} is less than \code{as.integer(lag)+2} an error will occur.}
}
\value{
A length \code{N+1} numeric vector, representing time steps 0 to N,
populated with zeros except at the time step (or steps if lag is a decimal)
representing the fixed delay from the origin.
}
\description{
Function to generate a linear response function (filter) representing a
fixed lag, including decimal time step lags.
}
\details{
For decimal lags, the mass of the response is shared between the
relevant consecutive time steps, in proportion to the fractional part of the
decimal. The first time step is assumed to represent a lag of zero.

For example, a lag of 1.2 would be shared between time step 2 and 3, with
respective masses, 0.8 and 0.2.
}
\examples{
\dontrun{
h <- make_lag_filter(1.2, 2)
}
}
