% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojstockr.R
\name{mojstockr_mconv}
\alias{mojstockr_mconv}
\title{Convolution of multiple time series with a matched set of filters.}
\usage{
mojstockr_mconv(x, h, non_data_cols = NULL)
}
\arguments{
\item{x}{A set of time series in 'horizontal' format with identifying, 'non-
data' columns on the left and data columns on the right. Data columns
contain the time series, one time series per row.}

\item{h}{A set of linear response functions, one for each time series. Non-
data columns are assumed to be the same as for \code{x}. Data columns are
assumed to start at zero lag.}

\item{non_data_cols}{A character vector of non-data columns to be excluded
from the convolution calculation.}
}
\value{
A tibble in the same format and dimensions as \code{x}, providing the
convolution of each row of \code{x} with the corresponding row of \code{h}.
}
\description{
Takes a set of time series, \code{x}, and linear response function (filter
kernel), \code{h}, and generates an output, \code{y}, as the convolution of
each time series with its filter.
}
\details{
Both \code{x} and \code{h} must be in a 'horizontal' format where, following
some non-data columns, consecutive time steps are represented as consecutive
columns. The non-data columns would typically identify the time series for a
given row. For example,

\code{h <- dataframe(sex = c("female", "male"), t0 = c(1, 1), t1 = c(0.6,
               0.5), t2 = c(0.3, 0.1))}

The first time step of the filter should be a the zero-lag response \emph{h[0]}.

The identity of any non-data columns should be specified as a character
vector of column names, e.g. \code{c("age", "sex")}

h must have the same number of rows as x but is not expected to have the same
number of data columns.

Convolution is with the base function, \code{stats::convolve()}. This uses
the Fast Fourier Transform and will return real-valued output. You may wish
to round output, if working with integer-valued systems.
}
\examples{
x <- data.frame(sex = c("female", "male"), t0 = rpois(2, 10),
               t1 = rpois(2, 10), t2 = rpois(2, 10), t3 = rpois(2, 10),
               t4 = rpois(2, 10), t5 = rpois(2, 10))
h <- data.frame(sex = c("female", "male"), lag0 = c(1, 1),
               lag1 = c(0.6, 0.5), lag2 = c(0.3, 0.1))
y <- mojstockr_mconv(x, h, non_data_cols = c("sex"))
}
