% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojstockr.R
\name{mojstockr_resample_profile}
\alias{mojstockr_resample_profile}
\title{Resample a discrete probability distribution that is defined at equidistant
time-steps (we refer to such a distribution as a profile)}
\usage{
mojstockr_resample_profile(
  x,
  probabilities,
  x_resampled = NULL,
  stretch_factor = NULL
)
}
\arguments{
\item{x}{Vector. Array containing points at which probabilities are defined.}

\item{probabilities}{Vector. Array containing probabilities corresponding to
the points defined in \code{x}.}

\item{x_resampled}{\emph{Optional.} Vector. Array containing points at which
resampled probabilities are defined.}

\item{stretch_factor}{\emph{Optional.} Float. The resampled probability
distribution will be stretched along the x-axis by the factor
\code{stretch_factor}.}
}
\value{
A tibble containing the resampled probability distribution, with
columns \code{x_resampled} and \code{probabilities_resampled}. The values
of \code{x_resampled} are separated by the same interval that separates
values of \code{x}.
}
\description{
\code{mojstockr_resample_profiles} uses linear interpolation to increase or
decrease the number of points at which a profile is defined. The input
profile consists of probabilities, \code{probabilities}, that are defined at
the equally spaced values \code{x}. One (but not both) of the optional
arguments \code{x_resampled} and \code{stretch_factor} must be supplied in
order to define the points at which resampled probabilities are calculated.
The values within \code{x_resampled} need not be monotonically increasing or
equally spaced. If \code{stretch_factor} is supplied,
\code{mojstockr_resample_profiles} generates a vector \code{x_resampled}
consisting of \code{ceiling(length(x) \* stretch_factor)} equally spaced
values that lie between \code{min(x \* stretch_factor)} and \code{max(x \*
stretch_factor)}.
}
\examples{
\dontrun{
dummy_profile <- data.frame(a = c("t0", "t1", "t2", "t3", "t4", "t5", "t6",
                                  "t7", "t8", "t9", "t10", "t11"),
                            t = c(0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11),
                            proportions = c(0.1, 0.2, 0.03, 0.07, 0.05, 0.05,
                                            0.01, 0.1, 0.1, 0.1, 0.1, 0.09)
)

returned_tibble <-
  mojstockr_resample_profiles(dummy_profile$t,
                              dummy_profile$proportions,
                              x_resampled = seq(from = 0, to = 24))

returned_tibble <-
  mojstockr_resample_profiles(dummy_profile$t, dummy_profile$proportions, 
                              stretch_factor = 2.5)
}
}
