% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_baseline}
\alias{run_baseline}
\title{A baseline run without any levers.}
\usage{
run_baseline(
  params,
  inflows_det_loaded,
  profiles_det_loaded,
  nomis_out_delius_in_ratio,
  profiles_lic,
  recall_rate_exclPSS,
  profiles_recall
)
}
\description{
A baseline run without any levers.
}
