% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_prison_determinate_module}
\alias{run_prison_determinate_module}
\title{Create a determinate sentence population from prison inflows and outflow
profiles.}
\usage{
run_prison_determinate_module(
  inflows_det,
  lever_profiles_det_stretch_impact_date,
  profiles_det_levered
)
}
\arguments{
\item{inflows_det}{A table of determinate sentence inflow volumes.}

\item{lever_profiles_det_stretch_impact_date}{Impact dates for applying
alternative determinate sentence time-served profiles.}

\item{profiles_det_levered}{Determinate sentence time-served profiles before
and after the impact date.}
}
\value{
A list including the outflows by determinate sentence length band and
their populations.
}
\description{
Exported only for use by the Shiny app. Not to be used by a user of the
package.
}
