% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_prison_indeterminate_module}
\alias{run_prison_indeterminate_module}
\title{Create an indeterminate sentence population from prison inflows.}
\usage{
run_prison_indeterminate_module(month_names, inflows_indet_mean)
}
\arguments{
\item{month_names}{The months defining the forecast period}

\item{inflows_indet_mean}{The mean indeterminate inflow rate (to be taken
from publihed statistics).}
}
\value{
A table representing the indeterminate population forecast.
}
\description{
Exported only for use by the Shiny app. Not to be used by a user of the
package.
}
