% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_prison_recall_module}
\alias{run_prison_recall_module}
\title{Create a recall prison population from estimates of the licence population.}
\usage{
run_prison_recall_module(
  outflows_det,
  nomis_out_delius_in_ratio,
  profiles_lic,
  recall_rate_exclPSS,
  lever_recall_rate,
  lever_recall_rate_impact_date,
  profiles_recall_levered,
  lever_profiles_recall_stretch_impact_date
)
}
\arguments{
\item{outflows_det}{The table of outflows by determinate sentence length
band.}

\item{nomis_out_delius_in_ratio}{Ratios, by sentence length band, for
converting prison outflows to typical probation inflows.}

\item{profiles_lic}{Licensee time-served profiles.}

\item{recall_rate_exclPSS}{Recall rate for those not on post-sentence
supervision (PSS).}

\item{lever_recall_rate}{The recall rates by sentence length band.}

\item{lever_recall_rate_impact_date}{The impact date for changing recall
rates.}

\item{profiles_recall_levered}{The time on recall profiles by sentence length
band,}

\item{lever_profiles_recall_stretch_impact_date}{The impact date for applying
the new recall profils}
}
\value{
A table representing the recall population forecast.
}
\description{
Exported only for use by the Shiny app. Not to be used by a user of the
package.
}
