% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_prisonsreadyreckoner}
\alias{run_prisonsreadyreckoner}
\title{Run \code{prisonsreadyreckoner} in its entirety.}
\usage{
run_prisonsreadyreckoner(params)
}
\description{
Run \code{prisonsreadyreckoner} in its entirety.
}
