% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{silence_botor}
\alias{silence_botor}
\title{Temporarily turn off debug messages when botor interacts with S3 while the
calling function is active.}
\usage{
silence_botor()
}
\description{
Temporarily turn off debug messages when botor interacts with S3 while the
calling function is active.
}
