% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prison.R
\name{split_populations_by_gender}
\alias{split_populations_by_gender}
\title{Split output populations by gender.}
\usage{
split_populations_by_gender(pop, gender_splits)
}
\arguments{
\item{pop}{A population table split by case type and sentence length band.}

\item{gender_splits}{A table of gender splits for each case type and
sentence length band.}
}
\value{
The population table with an additional column for gender, volumes
being pro-rated by relevant gender splits.
}
\description{
Exported only for use by the Shiny app. Not to be used by a user of the
package.
}
