% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levers.R
\name{stretch_profiles}
\alias{stretch_profiles}
\title{Stretch profiles by a user-defined factor}
\usage{
stretch_profiles(
  profiles,
  lever_profiles_det_stretch_factors,
  min_stretch_factor,
  max_stretch_factor,
  projection_length_months,
  non_data_cols = c("senband", "phase")
)
}
\description{
Stretch profiles by a user-defined factor
}
