% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levers.R
\name{stretch_recall_time_lever}
\alias{stretch_recall_time_lever}
\title{Stretch time on recall by a user-defined factor}
\usage{
stretch_recall_time_lever(
  recall_time,
  lever_profiles_recall_stretch_factors,
  min_stretch_factor,
  max_stretch_factor
)
}
\arguments{
\item{recall_time}{The recall times by sentence length band before
stretching.}

\item{lever_profiles_recall_stretch_factors}{The factors by sentence length
band by which to strech profiles.}

\item{min_stretch_factor}{A minimum allowed stretch factor.}

\item{max_stretch_factor}{A maximum allowed stretch factor.}
}
\value{
The new recall profiles with the post-impact profile stretched, as
requested.
}
\description{
Exported only for use by the Shiny app. Not to be used by a user of the
package.
}
