% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levers.R
\name{add_cc_sitting_days}
\alias{add_cc_sitting_days}
\title{Increase Crown Court capacity}
\usage{
add_cc_sitting_days(
  cc_capacity,
  lever_extra_cc_sitting_days,
  capacity_impact_date
)
}
\description{
Increase Crown Court capacity
}
