% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levers.R
\name{add_inflows_det_delta_lever}
\alias{add_inflows_det_delta_lever}
\title{Change number of prison receptions}
\usage{
add_inflows_det_delta_lever(
  inflows_det,
  lever_extra_inflows_det,
  lever_extra_inflows_det_impact_date,
  non_data_cols = "senband"
)
}
\description{
Change number of prison receptions
}
