% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_courts_module}
\alias{run_courts_module}
\title{Run Crown Court and magistrates' court modelling to generate change in prison
inflows owing to changes in the number of court disposals.}
\usage{
run_courts_module(
  cc_output,
  cc_capacity,
  cc_receipts_delta,
  mc_disposals_delta,
  profiles_remand_in,
  profiles_remand_out,
  sentencing_rates,
  inflows_det
)
}
\description{
Run Crown Court and magistrates' court modelling to generate change in prison
inflows owing to changes in the number of court disposals.
}
