% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{silence_botor}
\alias{silence_botor}
\title{Temporarily turn off debug messages when botor interacts with S3 while the
calling function is active.}
\usage{
silence_botor()
}
\value{
Not applicable.
}
\description{
Exported only for use by the Shiny app. Not to be used by a user of the
package.
}
