% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levers.R
\name{stretch_recall_time_lever}
\alias{stretch_recall_time_lever}
\title{Stretch time on recall by a user-defined factor}
\usage{
stretch_recall_time_lever(
  recall_time,
  lever_profiles_recall_stretch_factors,
  min_stretch_factor,
  max_stretch_factor
)
}
\description{
Stretch time on recall by a user-defined factor
}
